src/azure-ai-foundry/foundry-agentic-mesh/apps/agent-host-python/requirements.txt
azure-ai-projects>=1.0.0
azure-identity>=1.17.0
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
